fixed delete button

+++ routes/list-api-routes.js
@@ -31,11 +31,29 @@ module.exports = (app) => {
 
   // api call to delete (delete) one record in the list_members db
   app.delete("/api/lists/:id", (req, res) => {
-    db.ListMember.destroy({
+    db.GiftItem.findAll({
       where: {
-        id: req.params.id,
-      },
-    }).then((dbListMember) => res.json(dbListMember));
+        ListMemberId: req.params.id
+      }
+    }).then((dbGiftItems) => {
+        dbGiftItems.forEach( (GiftItem, index) => {
+            db.GiftItem.destroy({
+              where: {
+                id: GiftItem.id
+              }
+            }).then((dbGiftItem) => {
+              if(index === dbGiftItems.length - 1){
+                db.ListMember.destroy({
+                  where: {
+                    id: req.params.id,
+                  },
+                }).then((dbListMember) => res.json(dbListMember));
+              }
+              res.json(dbGiftItem)
+            })
+        })
+    })
+
 
   });
 };
